Extract app wrapper
This extracts everything related to setting up a single-page II app
(preloading some assets, setting up polyfills, compatibility checks etc)
into a helper function.

This is preparation work for adding new endpoints, so that they won't
need to import/use the main app code (more readable + smaller bundlers).

An apparently unnecessary endpoint (`#compatibilityNotice`) was also
removed.
